Fix for 2 battery acpi results
When executing the acpi command, it can result in multiple battery outputs. e.g.:

```
Battery 0: Discharging, 90%, 02:06:36 remaining
Battery 1: Discharging, 0%, rate information unavailable
```

In this case the return value will have two lines, resulting in only the percentage of the second battery shown, without bat_label or bat_stat.
Therefore, the fix will sort the `acpi` output based on the second column (percentage value) and sort it descending based on the values. Furthermore, it will cut the output of acpi after one line, so that all the information for the (hopefully) real battery will be shown.

Might have some bug, when the real battery has a 0% charge, but that should shutdown the computer anyway.